TubeDist: Reimplemented algorithm using AVX acceleration. Except this, also removed unnecessary mutex allocation in Processor.

+++ dafx/tubedist.hpp
@@ -42,8 +42,8 @@ public:
 	void process_params() override;
 	void process_callback() override;
 private:
-	float tC = 0, bC = 0, pC = 0;
-	float pre_gain = 0, gain = 0, out_gain = 0;
+	float8 tC = 0, bC = 0, pC = 0;
+	float8 pre_gain = 0, gain = 0, out_gain = 0;
 };
 }
 }
